+++ huanshankeji-team-gradle-plugins/src/main/kotlin/com/huanshankeji/team/gitversioning/MavenRepositoryContentFiltering.kt
@@ -1,0 +1,68 @@
+package com.huanshankeji.team.gitversioning
+
+import com.huanshankeji.gitversioning.DEV_COMMIT_VERSION_REGEX
+import com.huanshankeji.gitversioning.DIRTY_DEV_COMMIT_VERSION_REGEX
+import com.huanshankeji.gitversioning.LEGACY_SNAPSHOT_VERSION_REGEX
+import com.huanshankeji.team.HUANSHANKEJI_GROUP
+import org.gradle.api.artifacts.repositories.MavenArtifactRepository
+
+// TODO extract functions for non-`HUANSHANKEJI_MAVEN_GROUP` usages
+
+
+// TODO consider inlining and removing these `contentInclude` APIs
+
+fun MavenArtifactRepository.contentIncludeGroupVersions(
+    groupRegex: String,
+    versionRegex: String,
+) {
+    content {
+        includeVersionByRegex(groupRegex, ".*", versionRegex)
+    }
+}
+
+fun MavenArtifactRepository.contentExcludeGroupVersions(
+    groupRegex: String,
+    versionRegex: String,
+) {
+    content {
+        excludeVersionByRegex(groupRegex, ".*", versionRegex)
+    }
+}
+
+fun MavenArtifactRepository.contentIncludeModuleVersions(
+    groupRegex: String,
+    moduleRegex: String,
+    versionRegex: String,
+) {
+    content {
+        includeVersionByRegex(groupRegex, moduleRegex, versionRegex)
+    }
+}
+
+fun MavenArtifactRepository.contentExcludeModuleVersions(
+    groupRegex: String,
+    moduleRegex: String,
+    versionRegex: String,
+) {
+    content {
+        excludeVersionByRegex(groupRegex, moduleRegex, versionRegex)
+    }
+}
+
+
+const val HUANSHANKEJI_MAVEN_GROUP = HUANSHANKEJI_GROUP
+
+fun MavenArtifactRepository.contentIncludeHuanshankejiDirtyAndLegacySnapshots() {
+    contentIncludeGroupVersions(HUANSHANKEJI_MAVEN_GROUP, DIRTY_DEV_COMMIT_VERSION_REGEX)
+    contentIncludeGroupVersions(HUANSHANKEJI_MAVEN_GROUP, LEGACY_SNAPSHOT_VERSION_REGEX)
+}
+
+fun MavenArtifactRepository.contentIncludeHuanshankejiDevCommitVersions() {
+    contentIncludeGroupVersions(HUANSHANKEJI_MAVEN_GROUP, DEV_COMMIT_VERSION_REGEX)
+}
+
+fun MavenArtifactRepository.contentExcludeHuanshankejiNonStableVersions() {
+    contentExcludeGroupVersions(HUANSHANKEJI_MAVEN_GROUP, DEV_COMMIT_VERSION_REGEX)
+    contentExcludeGroupVersions(HUANSHANKEJI_MAVEN_GROUP, DIRTY_DEV_COMMIT_VERSION_REGEX)
+    contentExcludeGroupVersions(HUANSHANKEJI_MAVEN_GROUP, LEGACY_SNAPSHOT_VERSION_REGEX)
+}