+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/GithubPackagesMavenRegistry.kt
@@ -1,49 +1,38 @@
 package com.huanshankeji
 
-import org.gradle.api.Action
+import com.huanshankeji.github.packages.maven.githubPackagesMavenRegistry
+import com.huanshankeji.github.packages.maven.githubPackagesMavenRegistryWithName
+import com.huanshankeji.github.packages.maven.githubPackagesSetUrlAndCredentials
+import com.huanshankeji.publish.publishing
 import org.gradle.api.Project
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
-import org.gradle.api.publish.PublishingExtension
 import org.gradle.kotlin.dsl.repositories
 
-// TODO: use context receivers when it's stable.
+// All APIs in this file are deprecated and this file can be removed directly in the future.
+
+private const val GITHUB_PACKAGES_MAVEN_REGISTRY_OLD_APIS_DEPRECATION_MESSAGE =
+    "Use the new APIs in `com.huanshankeji.github.packages.maven` instead."
+
+@Deprecated(GITHUB_PACKAGES_MAVEN_REGISTRY_OLD_APIS_DEPRECATION_MESSAGE)
 fun Project.githubPackagesMavenRegistrySetUrlAndCredentials(
     mavenArtifactRepository: MavenArtifactRepository, owner: String, repository: String
 ) =
-    mavenArtifactRepository.run {
-        url = uri("https://maven.pkg.github.com/$owner/$repository")
-        credentials {
-            username = project.findProperty("gpr.user") as String? ?: System.getenv("USERNAME")
-            password = project.findProperty("gpr.key") as String? ?: System.getenv("TOKEN")
-        }
-    }
+    mavenArtifactRepository.githubPackagesSetUrlAndCredentials(owner, repository)
 
-// TODO: use context receivers when it's stable.
+@Deprecated(GITHUB_PACKAGES_MAVEN_REGISTRY_OLD_APIS_DEPRECATION_MESSAGE)
 fun Project.repositoriesAddGithubPackagesMavenRegistry(owner: String, repository: String) =
     repositories {
-        maven {
-            githubPackagesMavenRegistrySetUrlAndCredentials(this@maven, owner, repository)
-        }
+        githubPackagesMavenRegistry(owner, repository)
     }
 
-
+@Deprecated(GITHUB_PACKAGES_MAVEN_REGISTRY_OLD_APIS_DEPRECATION_MESSAGE)
 fun Project.publishingRepositoriesAddGithubPackagesMavenRepository(
-    nameArg: String = "GitHubPackages",
+    name: String = "GitHubPackages",
     owner: String,
-    repository: String
+    repository: String,
 ) =
-    // Copied and adapted from https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-gradle-registry.
     publishing {
         repositories {
-            maven {
-                name = nameArg
-                githubPackagesMavenRegistrySetUrlAndCredentials(this@maven, owner, repository)
-            }
+            githubPackagesMavenRegistryWithName(owner, repository, name)
         }
     }
-
-// copied and adapted from generated code
-fun Project.publishing(configure: Action<PublishingExtension>): Unit =
-    extensions.configure("publishing", configure)
-
-