+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/GradleCommonExperimentalApi.kt
@@ -1,0 +1,5 @@
+package com.huanshankeji
+
+@RequiresOptIn(message = "This API is experimental. It could change in the future without notice.")
+@Retention(AnnotationRetention.BINARY)
+annotation class GradleCommonExperimentalApi